requirements/base.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
install==1.3.5
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
astroid==2.9.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==21.1
importlib-metadata==4.10.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
toml==0.10.2
typing_extensions==4.0.1
wrapt==1.13.3
zipp==3.7.0

requests
